requirements.txt
torch>=1.10.0
torchvision
torch-geometric
pyyaml
opencv-python
numpy
scikit-learn
tensorboard
ultralytics
torchmetrics
